fix(review): move ProviderOutcome back to _type_results.py, fix __all__ ordering
Move ProviderOutcome from _type_results_execution.py back to _type_results.py
to eliminate the cross-import that undermined the test cascade narrowing.
SkillResult.provider references ProviderOutcome, and SkillResult is consumed
by 13+ directories — the cross-import meant changes to _type_results_execution
could break all those consumers but only 4 directories were tested.

Also swap _results_all and _results_execution_all ordering in types/__init__.py
__all__ concatenation to match import ordering.
