+++ tests/core/test_type_results_execution.py
@@ -13,11 +13,6 @@ class TestExecutionTypesImport:
 
         assert hasattr(SessionTelemetry, "empty")
 
-    def test_provider_outcome_importable(self):
-        from autoskillit.core.types._type_results_execution import ProviderOutcome
-
-        assert hasattr(ProviderOutcome, "none_used")
-
     def test_recipe_identity_importable(self):
         from autoskillit.core.types._type_results_execution import RecipeIdentity
 
@@ -32,15 +27,11 @@ class TestExecutionTypesImport:
         """All moved types are still importable from autoskillit.core."""
         from autoskillit.core import (
             CIRunScope,
-            ProviderOutcome,
             RecipeIdentity,
             SessionTelemetry,
         )
 
-        assert all(
-            cls is not None
-            for cls in [SessionTelemetry, ProviderOutcome, RecipeIdentity, CIRunScope]
-        )
+        assert all(cls is not None for cls in [SessionTelemetry, RecipeIdentity, CIRunScope])
 
 
 class TestExecutionTypesNotInResults:
@@ -49,20 +40,25 @@ class TestExecutionTypesNotInResults:
     def test_moved_types_absent_from_results_all(self):
         from autoskillit.core.types._type_results import __all__ as results_all
 
-        moved = {"SessionTelemetry", "ProviderOutcome", "RecipeIdentity", "CIRunScope"}
+        moved = {"SessionTelemetry", "RecipeIdentity", "CIRunScope"}
         overlap = moved & set(results_all)
         assert not overlap, f"Types still in _type_results.__all__: {overlap}"
+
+    def test_provider_outcome_in_results_all(self):
+        from autoskillit.core.types._type_results import __all__ as results_all
+
+        assert "ProviderOutcome" in results_all
 
     def test_present_in_execution_all(self):
         from autoskillit.core.types._type_results_execution import (
             __all__ as exec_all,
         )
 
-        expected = {"SessionTelemetry", "ProviderOutcome", "RecipeIdentity", "CIRunScope"}
+        expected = {"SessionTelemetry", "RecipeIdentity", "CIRunScope"}
         assert expected == set(exec_all)
 
     def test_skill_result_still_uses_provider_outcome(self):
-        """SkillResult.provider field default_factory references ProviderOutcome across modules."""
+        """SkillResult.provider field default_factory references ProviderOutcome."""
         from autoskillit.core import ProviderOutcome, SkillResult
 
         sr = SkillResult.crashed(Exception("test"))
