+++ tests/execution/test_linux_tracing_pty_integration.py
@@ -99,8 +99,8 @@ async def test_pty_wrapped_tracing_produces_no_script_snapshots_in_proc_trace_js
     Test 1.10 (partial): after the fix, every row in proc_trace.jsonl self-identifies
     the tracked process, and 'script' must never appear there.
     """
+    from autoskillit.core.types._type_results import ProviderOutcome
     from autoskillit.core.types._type_results_execution import (
-        ProviderOutcome,
         RecipeIdentity,
         SessionTelemetry,
     )
