+++ tests/execution/test_session_log_fields.py
@@ -10,8 +10,8 @@ from datetime import UTC, datetime
 
 import pytest
 
+from autoskillit.core.types._type_results import ProviderOutcome
 from autoskillit.core.types._type_results_execution import (
-    ProviderOutcome,
     RecipeIdentity,
     SessionTelemetry,
 )
